Major Bug fix (Scales and camera)
Removed unnecessary scaling factors in calculations. To achieve the same visual scale, I changed the default star_img_scale to 5000 ,img_scale to 100000 and distance scale to 100.

Bug fix: the position of new object when camera pos != (0,0)  used to be wrong due to the aforementioned scaling factors. The problem is now fixed.


Former-commit-id: 4d246b4b5f61e15358a167ff08c482a54afd4212